api/requirements.txt
requests
bs4
selenium
webdriver-manager
openai

flask
flasgger
Flask-SQLAlchemy
Flask-Migrate

psycopg2-binary

python-dotenv
